+++ test/src/test.rs
@@ -1,4 +1,5 @@
 mod arg_matcher_test;
+mod pickable_test;
 mod basic_test;
 mod multi_arg_test;
 mod multi_value_test;